ehn: Market data simulator with sine curve price instead of random
The sine curve will give a bit more realistic OHLC graph than before.
The buys and sells are now also predictable, making the tool more suitable for integration testing. 
The tool is now also self funding through the faucet, making it possible to run it for any length of time